+++ database/migrations/2017_01_01_108000_create_owners_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateOwnersTable extends Migration
+{
+    public function up()
+    {
+        Schema::create('owners', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('name')->unique();
+            $table->string('description')->nullable();
+            $table->boolean('is_active');
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('owners');
+    }
+}