+++ database/migrations/2017_01_01_111000_create_users_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateUsersTable extends Migration
+{
+    public function up()
+    {
+        Schema::create('users', function (Blueprint $table) {
+            $table->increments('id');
+            $table->integer('owner_id')->unsigned()->index('owner_id');
+            $table->foreign('owner_id')->references('id')->on('owners');
+            $table->integer('role_id')->unsigned()->index('roles_id');
+            $table->foreign('role_id')->references('id')->on('roles');
+            $table->string('email');
+            $table->string('password')->nullable();
+            $table->string('first_name')->nullable();
+            $table->string('last_name')->nullable();
+            $table->string('phone')->nullable();
+            $table->boolean('is_active');
+            $table->rememberToken();
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('users');
+    }
+}